+++ src/de/hhu/bsinfo/net/core/MessageImporterUnderOverflow.java
@@ -207,7 +207,7 @@ class MessageImporterUnderOverflow extends AbstractMessageImporter {
                 }
 
                 // read little endian byte order to big endian
-                ret |= (UnsafeMemory.readByte(m_bufferAddress + m_currentPosition) & 0xFF) << i * 8;
+                ret |= (long) (UnsafeMemory.readByte(m_bufferAddress + m_currentPosition) & 0xFF) << i * 8;
                 m_currentPosition++;
             }
             m_skippedBytes = m_skipBytes;
@@ -224,7 +224,7 @@ class MessageImporterUnderOverflow extends AbstractMessageImporter {
                 }
 
                 // read little endian byte order to big endian
-                ret |= (UnsafeMemory.readByte(m_bufferAddress + m_currentPosition) & 0xFF) << i * 8;
+                ret |= (long) (UnsafeMemory.readByte(m_bufferAddress + m_currentPosition) & 0xFF) << i * 8;
                 m_currentPosition++;
             }
             return ret;
